+++ examples/neli.rs
@@ -7,13 +7,13 @@ use neli::socket::*;
 fn main() {
     // Resolve generic netlink family ID
     let family_name = "your_family_name_here";
-    let mut sock = NlSocket::connect(NlFamily::Generic, None, None, true).unwrap();
+    let mut sock = NlSocket::connect(NlFamily::Generic, None, None).unwrap();
     let _id = sock.resolve_genl_family(family_name).unwrap();
 
     // Resolve generic netlink multicast group ID
     let family_name = "your_family_name_here";
     let group_name = "your_group_name_here";
-    let mut sock = NlSocket::connect(NlFamily::Generic, None, None, true).unwrap();
+    let mut sock = NlSocket::connect(NlFamily::Generic, None, None).unwrap();
     let _id = sock
         .resolve_nl_mcast_group(family_name, group_name)
         .unwrap();