+++ src/consts/nlattr.rs
@@ -24,3 +24,39 @@ impl_var_trait!(
     Name => libc::CTRL_ATTR_MCAST_GRP_NAME as u16,
     Id => libc::CTRL_ATTR_MCAST_GRP_ID as u16
 );
+
+/// Type representing attribute list types as indices
+#[derive(PartialEq)]
+pub struct Index(u16);
+
+impl Into<u16> for Index {
+    fn into(self) -> u16 {
+        self.0
+    }
+}
+
+impl From<u16> for Index {
+    fn from(v: u16) -> Self {
+        Index(v)
+    }
+}
+
+impl crate::Nl for Index {
+    fn serialize(&self, buf: &mut crate::StreamWriteBuffer) -> Result<(), crate::SerError> {
+        self.0.serialize(buf)?;
+        Ok(())
+    }
+
+    fn deserialize<B>(buf: &mut crate::StreamReadBuffer<B>) -> Result<Self, crate::DeError>
+    where
+        B: AsRef<[u8]>,
+    {
+        Ok(Index(u16::deserialize(buf)?))
+    }
+
+    fn size(&self) -> usize {
+        std::mem::size_of::<u16>()
+    }
+}
+
+impl NlAttrType for Index {}